Simplify pot expression
Since t = potl(min(64, lod_w, lod_h)), t is a power-of-two <= lod_w.
That implies that pot(aup(lod_w, t)) = pot(lod_w), so we can simplify
out the aup. (I wrote out the full proof in my notebook but it was
pretty messy and long, so not reproducing here.)

Signed-off-by: Asahi Yuni <yuni@asahilina.net>